# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Devian Hub API",
    "description": "\nAPI do hub do **Devian** — projetos, chats (Claude Code) e artefatos de build.\n\n**Autenticação:** os endpoints exigem um **Bearer token** com dois formatos:\n\n1. **Token de máquina** (`DEVIAN_API_TOKEN`) — automação/scripts; atua como o admin.\n2. **JWT de sessão** — retornado por `POST /auth/login` após login com a conta\n   Google (via `google_sign_in` no app; sem Firebase). O backend valida o ID\n   token contra as chaves públicas do Google e emite o token de sessão.\n\nOs dados (projetos, chats, mensagens e artefatos) são escopados pelo usuário\nautenticado: cada usuário vê apenas os próprios dados, resolvidos\nautomaticamente a partir do token.\n\nEndpoints públicos: `GET /health` e `POST /auth/login`.\nClique em **Authorize** e cole o token, sem o prefixo `Bearer `.\n",
    "version": "0.6.0"
  },
  "servers": [
    {
      "url": "https://api.agapech.com.br/devian",
      "description": "Produção — via túnel Cloudflare (oracle-hermi)"
    },
    {
      "url": "/devian",
      "description": "Local — backend direto na porta 8088"
    }
  ],
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user",
        "operationId": "get_me_users_me_get",
        "responses": {
          "200": {
            "description": "Perfil do usuário autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou sem usuário (token de máquina)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Conta inativa ou deletada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete account",
        "operationId": "delete_me_users_me_delete",
        "responses": {
          "204": {
            "description": "Inativa a conta do usuário (soft delete: status = 'deleted'). Login futuro com a mesma conta Google retorna 403."
          },
          "401": {
            "description": "Token ausente, inválido ou sem usuário (token de máquina)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Conta inativa ou deletada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user",
        "operationId": "update_me_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Atualiza nome, e-mail e/ou foto do usuário autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou sem usuário (token de máquina)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Conta inativa ou deletada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "E-mail já usado por outra conta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Corpo da requisição inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "description": "Lista de projetos (ordem alfabética)",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectOut"
                  },
                  "type": "array",
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create project",
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Projeto criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Já existe um projeto com esse nome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Corpo da requisição inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get project",
        "operationId": "get_project_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do projeto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Update project",
        "operationId": "update_project_projects__project_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projeto atualizado (PATCH-like: só os campos enviados)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Já existe um projeto com esse nome",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Corpo da requisição inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete project",
        "operationId": "delete_project_projects__project_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Projeto deletado (cascata: chats, mensagens e artefatos)"
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem de erro (em português)"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorOut",
        "description": "Resposta de erro padrão da API.",
        "example": {
          "message": "Projeto não encontrado"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Short project name. Becomes the unique identifier.",
            "example": "sisvisa"
          },
          "repo_url": {
            "title": "Repo Url",
            "description": "URL of the existing repository (git@ or https://).",
            "example": "git@bitbucket.org:branef/sisvisa-serr-mobile.git",
            "type": "string",
            "nullable": true
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Branch the project works on (used by builds and new chats).",
            "default": "main",
            "example": "develop"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "description": "Creates a project from an EXISTING repository.",
        "example": {
          "branch": "develop",
          "name": "sisvisa",
          "repo_url": "git@bitbucket.org:branef/sisvisa-serr-mobile.git"
        }
      },
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "title": "User Id",
            "description": "Owner user id. Null for projects created by machine token before the first login (adopted by the admin on login).",
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string",
            "nullable": true
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "repo_url",
          "branch",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectOut",
        "example": {
          "branch": "develop",
          "created_at": "2026-08-05T16:00:00Z",
          "id": "0191f3b2-4c3a-7b00-8000-000000000001",
          "name": "sisvisa",
          "repo_url": "git@bitbucket.org:branef/sisvisa-serr-mobile.git",
          "updated_at": "2026-08-05T16:00:00Z"
        }
      },
      "ProjectUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "nullable": true
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string",
            "nullable": true
          },
          "branch": {
            "title": "Branch",
            "type": "string",
            "nullable": true
          }
        },
        "type": "object",
        "title": "ProjectUpdate",
        "description": "Partial update — send only the fields you want to change.",
        "example": {
          "branch": "develop"
        }
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "picture_url": {
            "title": "Picture Url",
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "'admin' or 'member'"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "'active', 'blocked' or 'deleted'"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "picture_url",
          "role",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "UserOut",
        "example": {
          "created_at": "2026-08-05T16:00:00Z",
          "email": "samuel@agapech.com.br",
          "id": "0191f3b2-4c3a-7b00-8000-0000000000ff",
          "name": "Samuel",
          "picture_url": "https://lh3.googleusercontent.com/a/...",
          "role": "admin",
          "status": "active",
          "updated_at": "2026-08-05T16:00:00Z"
        }
      },
      "UserUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "nullable": true
          },
          "email": {
            "title": "Email",
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "nullable": true
          },
          "picture_url": {
            "title": "Picture Url",
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Partial update of the current user's profile.",
        "example": {
          "name": "Samuel G.",
          "picture_url": "https://lh3.googleusercontent.com/a/new"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Token Bearer do Devian — aceita dois formatos:\n1. Token de máquina (DEVIAN_API_TOKEN): atua como o admin (dono).\n2. JWT de sessão: retornado por POST /auth/login após login com Google.\nCole apenas o token, sem o prefixo 'Bearer '.",
        "scheme": "bearer"
      }
    }
  },
...
}